+++ crates/bcr-ebill-wasm/src/api/company.rs
@@ -45,8 +45,7 @@ impl Company {
         let file_bytes = get_ctx()
             .company_service
             .open_and_decrypt_file(id, file_name, &private_key)
-            .await
-            .map_err(|_| Error::NotFound)?;
+            .await?;
         get_ctx().contact_service.get_contact(id).await?; // check if contact exists
 
         let content_type = detect_content_type_for_bytes(&file_bytes)