+++ crates/bcr-ebill-core/src/bill/validation.rs
@@ -8,7 +8,7 @@ use crate::{
         },
     },
     constants::{ACCEPT_DEADLINE_SECONDS, PAYMENT_DEADLINE_SECONDS, RECOURSE_DEADLINE_SECONDS},
-    util,
+    util::{self, date::start_of_day_as_timestamp},
 };
 
 use super::{BillAction, BillIssueData, BillType, BillValidateActionData, RecourseReason};
@@ -34,8 +34,17 @@ pub fn validate_bill_issue(data: &BillIssueData) -> Result<(u64, BillType), Vali
         ));
     }
 
-    util::date::date_string_to_timestamp(&data.issue_date, None)?;
-    util::date::date_string_to_timestamp(&data.maturity_date, None)?;
+    let issue_date_ts = util::date::date_string_to_timestamp(&data.issue_date, None)?;
+    let maturity_date_ts = util::date::date_string_to_timestamp(&data.maturity_date, None)?;
+    let start_of_day = start_of_day_as_timestamp(data.timestamp);
+
+    if maturity_date_ts < start_of_day {
+        return Err(ValidationError::MaturityDateInThePast);
+    }
+
+    if issue_date_ts > maturity_date_ts {
+        return Err(ValidationError::IssueDateAfterMaturityDate);
+    }
 
     let bill_type = match data.t {
         0 => BillType::PromissoryNote,
@@ -591,8 +600,8 @@ mod tests {
             t: 0,
             country_of_issuing: "AT".into(),
             city_of_issuing: "Vienna".into(),
-            issue_date: "2024-08-12".into(),
-            maturity_date: "2024-11-12".into(),
+            issue_date: "2025-08-12".into(),
+            maturity_date: "2025-11-12".into(),
             drawee: TEST_PUB_KEY_SECP.into(),
             payee: OTHER_TEST_PUB_KEY_SECP.into(),
             sum: "500".into(),
@@ -618,6 +627,8 @@ mod tests {
     #[case::invalid_file_id( BillIssueData { file_upload_ids: vec!["".into()], ..valid_bill_issue_data() }, ValidationError::InvalidFileUploadId)]
     #[case::invalid_issue_date( BillIssueData { issue_date: "invaliddate".into(), ..valid_bill_issue_data() }, ValidationError::InvalidDate)]
     #[case::invalid_maturity_date( BillIssueData { maturity_date: "invaliddate".into(), ..valid_bill_issue_data() }, ValidationError::InvalidDate)]
+    #[case::maturity_date_before_now( BillIssueData { maturity_date: "2004-01-12".into(), ..valid_bill_issue_data() }, ValidationError::MaturityDateInThePast)]
+    #[case::issue_date_after_maturity_date( BillIssueData { issue_date: "2028-01-12".into(), ..valid_bill_issue_data() }, ValidationError::IssueDateAfterMaturityDate)]
     #[case::invalid_bill_type( BillIssueData { t: 5, ..valid_bill_issue_data() }, ValidationError::InvalidBillType)]
     #[case::drawee_equals_payee( BillIssueData { drawee: TEST_PUB_KEY_SECP.into(), payee: TEST_PUB_KEY_SECP.into(), ..valid_bill_issue_data() }, ValidationError::DraweeCantBePayee)]
     #[case::invalid_payee( BillIssueData { payee: "invalidkey".into(), ..valid_bill_issue_data() }, ValidationError::InvalidSecp256k1Key("invalidkey".into()))]